32bit subghz hashes (less false duplicates)

+++ applications/main/subghz/subghz_history.h
@@ -33,9 +33,9 @@ void subghz_history_delete_item(SubGhzHistory* instance, uint16_t idx);
  * 
  * @param instance - SubGhzHistory instance
  * @param idx      - record index
- * @return hash    - hash data byte
+ * @return hash    - hash data
  */
-uint8_t subghz_history_get_hash_data(SubGhzHistory* instance, uint16_t idx);
+uint32_t subghz_history_get_hash_data(SubGhzHistory* instance, uint16_t idx);
 
 /** Get repeat count to history[idx]
  *
